Fix -Wformat: use %lu for uint32_t in fprintf/coap_log calls
Three call sites passed uint32_t to %u format specifiers. On 32-bit ARM
(and other ABIs where 'unsigned int' is 16 bits or where uint32_t
typedefs to 'unsigned long'), this is a format/type mismatch and gcc
flags it with -Wformat= correctly. The behavior is fine in practice
when sizes happen to match but is UB per the C standard.

  - src/coap_subscribe.c: observe_num (uint32_t) printed with %u in
    three fprintf calls. Cast to (unsigned long) and use %lu.

  - src/coap_oscore.c: coap_opt_length() (uint32_t) used as %.*s
    precision specifier, which requires int. Cast to (int).

  - examples/coap-server.c: same coap_opt_length() %.*s issue.

Matches libcoap's existing style (the codebase uses %lu + cast
elsewhere rather than the PRIu32 macro from <inttypes.h>).

+++ examples/coap-server.c
@@ -1019,7 +1019,7 @@ hnd_proxy_uri(coap_resource_t *resource COAP_UNUSED,
   proxy_uri = coap_check_option(request, COAP_OPTION_PROXY_URI, &opt_iter);
   if (proxy_uri) {
     coap_log_info("Proxy URI '%.*s'\n",
-                  coap_opt_length(proxy_uri),
+                  (int)coap_opt_length(proxy_uri),
                   (const char *)coap_opt_value(proxy_uri));
     if (coap_split_proxy_uri(coap_opt_value(proxy_uri),
                              coap_opt_length(proxy_uri),
